# Simplify/optimize cached UID/GID lookups